+++ app/Http/Controllers/V1/Controller.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\V1;
 
 use Laravel\Lumen\Routing\Controller as BaseController;
